Simplify empty text box behaviour
Re-rendering the quotes depending on presence of text seemed to lead to re-running cached data load function. I don't know if that's a streamlit bug or my understanding is flaky. In any case, this fixes the problem, and gives somewhat appropriate quotes for an empty box. Also changed the default text to be more Les Mis.